requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-dotenv==1.0.0
langchain==0.0.173
langchain-openai==0.0.2
google-auth-oauthlib==1.0.0
google-api-python-client==2.86.0
msal==1.22.0
caldav==1.2.0
aiohttp==3.8.4
aiosqlite==0.19.0
python-multipart==0.0.6
openai-whisper==20231117
soundfile==0.12.1
numpy==1.24.3
